# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZGalaxy - Planet & Moon Infrastructure Management API
  description: |
    ZGalaxy Planet & Moon Infrastructure Engine API. Manages self-hosted ZeroTier
    Planet/Moon infrastructure: authentication & RBAC, planet/moon building,
    federation mesh, cluster HA, Cloudflare DNS/DDNS, identity, and encrypted backup.
  version: 1.3.0
servers:
  - url: http://localhost:3000
    description: Local ZGalaxy server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
paths:
...
  /api/v1/moons/create:
    post:
      summary: Create a signed Moon
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoints]
              properties:
                name: { type: string }
                endpoints:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Moon created
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/moons/{id}:
    put:
      summary: Update a moon's endpoints
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Moon updated
        '403':
          $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a moon file (ADMIN only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Moon deleted
        '400':
          description: Invalid moon identifier (path traversal rejected)
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/moons/{id}/migrate:
    post:
      summary: Migrate a moon to a target planet (identity-preserving)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetPlanetId]
              properties:
                targetPlanetId: { type: string }
      responses:
        '200':
          description: Moon migrated
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/moons/{id}/rebind:
    post:
      summary: Rebind a moon's endpoints
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoints]
              properties:
                endpoints:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Moon rebound
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/moons/{id}/rebuild:
    post:
      summary: Rebuild a moon's signed file
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Moon rebuilt
        '403':
          $ref: '#/components/schemas/Error'
  /api/v1/moons/{id}/download:
    get:
      summary: Download a moon file (authenticated)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Binary .moon file
        '400':
          description: Invalid moon identifier
...